02. DataSanitization/requirements.txt
artifacts-keyring==0.2.9
astroid==2.4.2
autopep8==1.5.4
azure-core==1.8.2
azure-cosmos==4.2.0
azure-functions==1.4.0
azure-functions-worker==1.1.6
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coverage==5.2
docutils==0.16
grpcio==1.26.0
grpcio-tools==1.26.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
joblib==0.17.0
keyring==21.2.1
lazy-object-proxy==1.4.3
linecache2==1.0.0
mccabe==0.6.1
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pip-check-reqs==2.1.1
pkginfo==1.5.0.1
protobuf==3.13.0
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.47.0
traceback2==1.4.0
twine==3.2.0
typed-ast==1.4.1
unittest2==1.1.0
urllib3==1.25.9
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
